Add support for setting the image registry globally
When we want to set a different registry than docker hub we have to do so like:

```yaml
image:
  repository: docker.exampleorg.com/rosette/server-enterprise
rootsImageRepository: docker.exampleorg.com/rosette
indoccoref:
  repository: docker.exampleorg.com/rosette/indoc-coref-server
```

This requires us to duplicate the registry in multiple places while having to remember the different suffixes for the different images. This change adds a global registry setting that can be used like this:

```yaml
global:
  imageRegistry: docker.exampleorg.com
```

This will set the registry for all images to `docker.exampleorg.com`.

This becomes more useful when this chart is being used as a dependency in another chart. When all dependency charts use `global.imageRegistry` the person installing the parent chart only has to set the registry once.

The `global.imageRegistry` setting is implemented in the most popular Helm chart registry and is considered a best-practice.

This change is non-breaking and backwards compatible. If `global.imageRegistry` is not set the chart will behave as before.

Signed-off-by: yardenshoham <git@yardenshoham.com>